==> Notebooks/Step_0_Grab_data.ipynb <==
# Step 0: Grab data files from NCES
# This page is just to grab the files we need

## At the core, each "survey" file has the csv data along with a dictionary for deciphering it
### These will be downloaded to a local "input directory"

import pandas as pd
import numpy as np
from zipfile import ZipFile
import csv

# This function is just a utility for grabbing the files
import urllib
import io
import os

def grab_zipfile(baseUrl, fileroot):
    '''uses a known static baseUrl to grab a zipped archive'''
    full_url = baseUrl + fileroot + '.zip'
    print(full_url, flush=True)
    remoteFile = urllib.request.urlopen(full_url)

    bio = io.BytesIO()
    bio.write(remoteFile.read())

    z = ZipFile(bio, 'r')
    print(z.namelist(), flush=True)
    z.extract(z.namelist()[0])
    return z.namelist()[0]

os.mkdir('inputs')
os.chdir('inputs')

# Here's where the files live:
baseUrl = 'https://nces.ed.gov/ipeds/datacenter/data/'

# Grad rate data; for each file, we typically try 2 more years than the prior year and if that file doesn't exist, try one more
fileroot = 'GR2021'

grab_zipfile(baseUrl, fileroot)
grab_zipfile(baseUrl, fileroot+'_DICT')


# For grad rate only, we grab the 3 most recent years
grab_zipfile(baseUrl, 'GR2020')
grab_zipfile(baseUrl, 'GR2019')

# The master directory:
grab_zipfile(baseUrl, 'HD2021')

grab_zipfile(baseUrl, 'HD2021_DICT')

# Admissions
grab_zipfile(baseUrl, 'ADM2021')
grab_zipfile(baseUrl, 'ADM2021_DICT')

# Ethnicity
grab_zipfile(baseUrl, 'EF2021A')
grab_zipfile(baseUrl, 'EF2021A_DICT')

# Retention
grab_zipfile(baseUrl, 'EF2021D')
grab_zipfile(baseUrl, 'EF2021D_DICT')

# Institutional Characteristics
grab_zipfile(baseUrl, 'IC2021')
grab_zipfile(baseUrl, 'IC2021_DICT')

# Student Financial Aid
grab_zipfile(baseUrl, 'SFA2021')
grab_zipfile(baseUrl, 'SFA2021_DICT')
